flytt refusjon status inn i user info
