+++ server/server.js
@@ -172,7 +172,6 @@ const main = async () => {
         (await import('./mock/antallArbeidsforholdMock.js')).mock(app);
         (await import('./mock/tiltakApiMock.js')).mock(app);
         (await import('./mock/sykefraværMock.js')).mock(app);
-        (await import('./mock/refusjonsStatusMock.js')).mock(app);
         (await import('./mock/presenterteKandidaterMock.js')).mock(app);
         (await import('./mock/storageMock.js')).mock(app);
         (await import('./mock/kontaktinfoApiMock.js')).mock(app);
