+++ src/api/dnaApi.ts
@@ -1,32 +1,3 @@
-import { z } from 'zod';
-import * as Sentry from '@sentry/browser';
-
-const RefusjonStatus = z.object({
-    virksomhetsnummer: z.string(),
-    statusoversikt: z.object({
-        KLAR_FOR_INNSENDING: z.number().optional(),
-    }),
-    tilgang: z.boolean(),
-});
-const RefusjonStatusResponse = z.array(RefusjonStatus);
-export type RefusjonStatus = z.infer<typeof RefusjonStatus>;
-
-const refusjonstatusURL = '/min-side-arbeidsgiver/api/refusjon_status';
-export async function hentRefusjonstatus(): Promise<RefusjonStatus[]> {
-    const respons = await fetch(refusjonstatusURL);
-    if (respons.ok) {
-        const data = await respons.json();
-        try {
-            return RefusjonStatusResponse.parse(data);
-        } catch (error) {
-            Sentry.captureException(error);
-        }
-    }
-    throw new Error(
-        `Kall til ${refusjonstatusURL} feilet med ${respons.status}:${respons.statusText}`
-    );
-}
-
 const sjekkInnloggetURL = '/min-side-arbeidsgiver/api/innlogget';
 export const sjekkInnlogget = async (): Promise<boolean> => {
     const { ok } = await fetch(sjekkInnloggetURL);
